requirements.txt
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.17.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
oauthlib==3.1.0
pyluach==1.1.1
python-hebrew-numbers==0.2.3